fix some broken imports + usage of old classes + remove migration package and move model into parent package of models
